requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-auditlog==0.4.7
django-crispy-forms==1.8.0
django-import-export==2.0
django-jsonfield==1.4.0
et-xmlfile==1.0.1
jdcal==1.4.1
lxml==4.4.2
MarkupPy==1.14
mysqlclient==1.4.4
numpy==1.17.4
odfpy==1.4.0
openpyxl==3.0.2
pandas==0.25.3
passlib==1.7.1
python-dateutil==2.6.0
python-docx==0.8.10
pytz==2019.3
PyYAML==5.2
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
wfastcgi==3.0.0
xlrd==1.2.0
xlwt==1.3.0
